services/eon-voice-claude/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
websockets>=12.0
azure-monitor-opentelemetry>=1.2.0
